// src/app/modules/academics/grading/models/test-score-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets TestScoreView
 */
export interface TestScoreView { 
    /**
     * Gets or Sets TestScheduleId
     */
    testScheduleId?: string;
    /**
     * Gets or Sets StudentId
     */
    studentId?: string;
    /**
     * Gets or Sets Score
     */
    score?: number;
}