// src/app/modules/academics/grading/models/studentScoreBulkView.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TestScoreView } from './test-score-view';
import { AttachmentViewModel } from '../../student-admissions/model/attachment-view-model';


/**
 * StudentScoreBulkView
 */
export interface StudentScoreBulkView {
    /**
     * TestScoreViews
     */
    testScoreViews?: Array<TestScoreView>;
    /**
     * Attachment ViewModel
     */
    attachmentViewModel?: AttachmentViewModel;
}